favorites, photos

+++ realtyStore/Controllers/RealtorModeController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
@@ -24,31 +24,57 @@ namespace realtyStore.Controllers
         {
             if (action == "showOwners")
             {
-                ViewBag.Table = db.Owners;
-                foreach (Owner o in ViewBag.Table)
+                IEnumerable<myUser> owners = new myUser[] { };
+                foreach(var r in db.Realties)
                 {
-                    o.Address = db.Cities.FirstOrDefault(c => c.Id == o.CityId).Name + ' ' + o.Address;
+                    var owner = db.Users.FirstOrDefault(u => u.Id == r.OwnerId);
+                    owner.Address = "г. " + db.Cities.FirstOrDefault(c => c.Id == owner.CityId).Name + ", " + owner.Address;
+                    if (owners.FirstOrDefault(b => b.Id == owner.Id) == null)
+                    {
+                        owners = owners.Prepend(owner);
+                    }
                 }
-                ViewBag.Count = db.Owners.Count();
+                   
+                ViewBag.Table = owners.Reverse();
+
 
                 ViewBag.TableName = "Владельцы";
+                ViewBag.Count = owners.ToList().Count;
             }
-            if (action == "showBuyers")
+            else if (action == "showBuyers")
             {
                 ViewBag.TableName = "Арендаторы/покупатели";
-                ViewBag.Count = db.Buyers.Count();
-
-                ViewBag.Table = db.Buyers;
-                foreach (Buyer b in ViewBag.Table)
+                IEnumerable<myUser> buyers = new myUser[] { };
+                foreach (var l in db.Leased)
                 {
-                    b.Address = "г." + db.Cities.FirstOrDefault(c => c.Id == b.CityId).Name + ", " + b.Address;
+                    var buyer = db.Users.FirstOrDefault(u => u.Id == l.BuyerId);
+                    buyer.Address = "г. " + db.Cities.FirstOrDefault(c => c.Id == buyer.CityId).Name + ", " + buyer.Address;
+                    if (buyers.FirstOrDefault(b => b.Id == buyer.Id) == null)
+                    {
+                        buyers = buyers.Prepend(buyer);
+                    }
                 }
+                
+                foreach (var s in db.Sold)
+                {
+                    var buyer = db.Users.FirstOrDefault(u => u.Id == s.BuyerId);
+                    buyer.Address = "г. " + db.Cities.FirstOrDefault(c => c.Id == buyer.CityId).Name + ", " + buyer.Address;
+                    if(buyers.FirstOrDefault(b => b.Id == buyer.Id) == null)
+                    {
+                        buyers = buyers.Prepend(buyer);
+                    }
+                }
+
+                ViewBag.Table = buyers.Reverse();
+
+                ViewBag.Count = buyers.ToList().Count;
             }
             return View();
         }
 
 
         [HttpGet]
+        [Authorize(Roles = "realtor")]
         public ActionResult Dial()
         {
             ViewBag.TableNameSold = "Сделки по продаже";
@@ -74,6 +100,7 @@ namespace realtyStore.Controllers
             return View();
         }
         [HttpPost]
+        [Authorize(Roles="realtor")]
         public ActionResult Dial(string action, string DateCheckout, int? Id)
         {
             if (action == "SaveNewDate")
@@ -86,13 +113,17 @@ namespace realtyStore.Controllers
                 
                 foreach (Leased l in db.Leased)
                 {
-                    var d1 = DateTime.Parse(l.DateCheckOut).Date;
-                    var d2 = DateTime.Now.Date;
-                    int result = DateTime.Compare(d1, d2);
-                    if (result < 0)
+                    if (l != null)
                     {
-                        db.Realties.FirstOrDefault(r => r.Id == l.RealtyId).Status = l.Status;
-                        db.Leased.Remove(l);
+                        var d1 = DateTime.Parse(l.DateCheckOut).Date;
+                        var d2 = DateTime.Now.Date;
+                        int result = DateTime.Compare(d1, d2);
+                        if (result < 0)
+                        {
+                            db.Realties.FirstOrDefault(r => r.Id == l.RealtyId).Status = l.Status;
+                            db.Leased.Remove(l);
+                        }
+
                     }
                 }
                 db.SaveChanges();
@@ -124,6 +155,7 @@ namespace realtyStore.Controllers
 
 
         [HttpGet]
+        [Authorize(Roles="realtor")]
         public ActionResult Apps()
         {
             ViewBag.TableName = "Заявки";           
@@ -137,58 +169,33 @@ namespace realtyStore.Controllers
             return View();
         }
 
-        public string Owner(int? id)
+        [Authorize(Roles="realtor")]
+        public ActionResult Realtor(int? id)
         {
-            Owner owner = db.Owners.FirstOrDefault(o => o.Id == id);
-            if (owner != null)
-            {
-                return owner.LastName + ' ' 
-                    + owner.FirstName + ' ' 
-                    + owner.Patronymic + ' ' 
-                    + owner.Passport + ' ' 
-                    + owner.Phone + ' ' 
-                    + db.Cities.FirstOrDefault(c => c.Id == owner.CityId).Name + ' ' + owner.Address;
-            }
-            return "Не найден!";
+            myUser realtor = db.Users.FirstOrDefault(r => r.Id == id);
+            ViewBag.men = realtor;
+            ViewBag.address = db.Cities.FirstOrDefault(c => c.Id == realtor.CityId).Name + ", " + realtor.Address;
+            return View();
         }
-        public string Realtor(int? id)
+        [Authorize(Roles="realtor")]
+        public ActionResult ShowUser(int? id)
         {
-            Realtor realtor = db.Realtors.FirstOrDefault(r => r.Id == id);
-            if (realtor != null)
-            {
-                return realtor.LastName + ' '
-                    + realtor.FirstName + ' '
-                    + realtor.Patronymic + ' '
-                    + realtor.Passport + ' '
-                    + realtor.Phone + ' '
-                    + db.Cities.FirstOrDefault(c => c.Id == realtor.CityId).Name + ' ' + realtor.Address;
-            }
-            return "Не найден!";
+            myUser user = db.Users.FirstOrDefault(b => b.Id == id);
+            ViewBag.men = user;
+            ViewBag.address = db.Cities.FirstOrDefault(c => c.Id == user.CityId).Name + ", " + user.Address;
+            return View();
         }
-        public string Buyer(int? id)
-        {
-            Buyer buyer = db.Buyers.FirstOrDefault(b => b.Id == id);
-            if (buyer != null)
-            {
-                return buyer.LastName + ' '
-                    + buyer.FirstName + ' '
-                    + buyer.Patronymic + ' '
-                    + buyer.Passport + ' '
-                    + buyer.Phone + ' '
-                    + db.Cities.FirstOrDefault(c => c.Id == buyer.CityId).Name + ' ' + buyer.Address;
-            }
-            return "Не найден!";
-        }
+        [Authorize(Roles="realtor")]
         public ActionResult PlaceApp(int? id)
         {
             ApplicationToRealtor app = db.Apps.FirstOrDefault(a => a.Id == id);
-            Owner owner = db.Owners.FirstOrDefault(a => a.Patronymic == app.Patronymic 
+            myUser owner = db.Users.FirstOrDefault(a => a.Patronymic == app.Patronymic 
                                                      && a.LastName == app.LastName 
                                                      && a.FirstName == app.FirstName 
                                                      || a.Phone == app.Phone);
             if (owner == null)
             {
-                owner = new Owner { CityId = app.CityId,
+                owner = new myUser { CityId = app.CityId,
                     FirstName = app.FirstName,
                     LastName = app.LastName,
                     Patronymic = app.Patronymic,
@@ -196,23 +203,24 @@ namespace realtyStore.Controllers
                     Phone = app.Phone,
                     Address = app.Address
                 };
-                db.Owners.Add(owner);
+                db.Users.Add(owner);
                 db.SaveChanges();
             }
 
             db.Realties.Add(new Realty{ ImgUrl = imgUrl, Type = app.RealtyType, NumberRoom = app.NumberRoom, Address = app.Address, Square = app.Square, Floor = app.Floor, Floors = app.Floors, Status = app.Status, CityId = app.CityId, OwnerId = owner.Id, Price = app.Price, Description = app.Description, RealtorId = 1 });
-            //db.SaveChanges();
             db.Apps.Remove(app);
             db.SaveChanges();
             return RedirectToAction("Apps", "RealtorMode");
         }
         [HttpGet]
+        [Authorize(Roles="realtor")]
         public ActionResult EditApp(int? id)
         {
             ViewBag.Cities = db.Cities;
             return View(db.Apps.FirstOrDefault(a => a.Id == id));
         }
         [HttpPost]
+        [Authorize(Roles="realtor")]
         public ActionResult EditApp(ApplicationToRealtor app)
         {
             ApplicationToRealtor oldApp = db.Apps.FirstOrDefault(a => a.Id == app.Id);
@@ -233,8 +241,9 @@ namespace realtyStore.Controllers
             db.SaveChanges();
             return RedirectToAction("Apps", "RealtorMode");
         }
-        [HttpGet]
 
+        [HttpGet]
+        [Authorize(Roles = "realtor")]
         public ActionResult Realties()
         {
             ViewBag.TableName = "Все объявления";
@@ -248,6 +257,7 @@ namespace realtyStore.Controllers
             return View();
         }
         [HttpPost]
+        [Authorize(Roles="realtor")]
         public ActionResult Realties(int? Id)
         {
             db.Realties.Remove(db.Realties.FirstOrDefault(r => r.Id == Id));
@@ -259,12 +269,14 @@ namespace realtyStore.Controllers
         }
        
         [HttpGet]
+        [Authorize(Roles="realtor")]
         public ActionResult EditRealty(int? id)
         {
             ViewBag.Cities = db.Cities;
             return View(db.Realties.FirstOrDefault(r => r.Id == id));
         }
         [HttpPost]
+        [Authorize(Roles="realtor")]
         public ActionResult EditRealty(int Id, string Status, string Description, string date)
         {
             Realty realty = db.Realties.FirstOrDefault(a => a.Id == Id);
@@ -302,7 +314,7 @@ namespace realtyStore.Controllers
                     Date = DateTime.Now.Date.ToString("yyyy-MM-dd"),
                     Price = realty.Price,
                     RealtorId = realty.RealtorId,
-                    OldOwnerId = realty.OwnerId,
+                    OwnerId = realty.OwnerId,
                 };
                 db.Sold.Add(s);
                 db.Realties.Remove(realty);
@@ -316,6 +328,7 @@ namespace realtyStore.Controllers
         }
 
         [HttpGet]
+        [Authorize(Roles="realtor")]
         public ActionResult BuyerForm(int? id, string type)
         {
             ViewBag.Cities = db.Cities;
@@ -324,23 +337,24 @@ namespace realtyStore.Controllers
             return View();
         }
         [HttpPost]
-        public ActionResult BuyerForm(int dialId, string type, Buyer buyer)
+        [Authorize(Roles="realtor")]
+        public ActionResult BuyerForm(int dialId, string type, myUser buyer)
         {
-            Buyer findBuyer = db.Buyers.FirstOrDefault(a => a.Patronymic == buyer.Patronymic && a.LastName == buyer.LastName && a.FirstName == buyer.FirstName || a.Phone == buyer.Phone);
+            myUser findBuyer = db.Users.FirstOrDefault(a => a.Patronymic == buyer.Patronymic && a.LastName == buyer.LastName && a.FirstName == buyer.FirstName || a.Phone == buyer.Phone);
             if (findBuyer == null)
             {
-                db.Buyers.Add(buyer);
+                db.Users.Add(buyer);
                 db.SaveChanges();
             }
             if (type == "sold")
             {
-                db.Sold.FirstOrDefault(s => s.Id == dialId).NewOwnerId = buyer.Id;
+                db.Sold.FirstOrDefault(s => s.Id == dialId).OwnerId = buyer.Id;
             }
             if (type == "leased")
             {
                 db.Leased.FirstOrDefault(l => l.Id == dialId).BuyerId = buyer.Id;
             }
-                db.SaveChanges();
+            db.SaveChanges();
             return RedirectToAction("Realties", "RealtorMode");
         }
     }
